requirements.txt
gradio
pandas
numpy
scikit-fuzzy
gitpython
requests
google-generativeai